Move remount/viewport helpers out of webPanel to satisfy max-lines
